Moving tokenizer out to its own file
